grid settings moved, but row height adjustment broken
